requirements.txt
tensorflow>=2.10
numpy
pandas
matplotlib
scikit-learn
opencv-python
pillow
seaborn
tqdm